move test_closed_group from sets to __init__
