+++ sympy/assumptions/handlers/__init__.py
@@ -110,3 +110,23 @@ class TautologicalHandler(AskHandler):
         if qt is None:
             return None
         return pt == qt
+
+
+#### Helper methods
+def test_closed_group(expr, assumptions, key):
+    """
+    Test for membership in a group with respect
+    to the current operation
+    """
+    result = True
+    for arg in expr.args:
+        _out = ask(key(arg), assumptions)
+        if _out is None:
+            break
+        elif _out is False:
+            if result:
+                result = False
+            else:
+                break
+    else:
+        return result
